Solution to the “Yesterday” exercise, as posted on the (unofficial) “C for Dummies” blog on 2023-05-01, implementing a 'calendar' module with functions such as calendar_today(), calendar_day_before() and calendar_day_after().
Now, while the Unix Epoch starts on January 1st, 1970 at 00:00:00 UTC,
on Windows the Win32 Epoch starts on January 1st, 1601 at 00:00:00 UTC.

Finding this a bit too limiting, the accompanying 'calendar' module
provides its own implementation of the Julian and Gregorian calendar;
thus allowing one to iterate over all dates starting from

[Julian calendar]    0004-03-01 C.E., JDN 1722578 all the way up to
[Gregorian calendar] 3440-12-31 C.E., JDN 2977859

That's because, while the Julian calendar took effect on 1 Jan. 45 BCE,
the pontifices initially added a leap day every 3 (instead of every 4)
years. This “Leap year error” was later corrected by Augustus, but the
upshot of all this is that the first (Julian) date whose day is known
with certainty is 1 March 4 CE (a Saturday).

On the other hand, the last date was chosen based on E. R. Hope's
Gregorian calendar drift function (as given in Journal of the Royal
Astronomical Societ of Canada, Vol. 58, 1964, p.4), which indicates
that curr. leap year rules will add up to an (as of now unscheduled)
intercalary day by the year 3441.