A good introduction to memory management algorithms in C can be found in Bill Blunden's classic book "Memory Management: Algorithms and Implementation in C/C++".
This adds a "Malloc in C.7z" archive containing sample implemen-
tations for many algorithms discussed in this book (and also adds
a few additional ones):

* Bitmapped Fit algorithm
* Implicit free lists
* Singly-linked (free) lists
* Doubly-linked (free) lists
* Segregated free lists
* Buddy allocator
* Arena allocator (i.e. Region-based memory management)
* A (3d-Point) sub-allocator

The included "main.c" tests these allocators by timing how long it
takes to perform a total of 7500 allocations+2500 free operations.
For comparison, the same (de)allocation requests are also passed
to the malloc() and free() functions of the ‘C’ standard library.
For completeness, they are in the end also passed to the malloc()
interface of the Boehm-Demers-Weiser Garbage Collector (v8.2.8).

Lastly, the results are presented to the user in form of a list
of (relative) performance numbers.