On Sun, 5 Jan 2025 Ken Pizzini announced the latest release of the Free Software Foundationʼs “bc” command-line calculator.
This is a Windows-compatible 64-bit build of this calculator,
compiled with Visual Studio 2022, Version 17.13.6.

"bc-functions\" contains collections of functions written in
the bc language (taken from http://phodd.net/gnu-bc/ and other
sites) that may be of interest. The accompanying "ebc.bat"
runs the "bc.exe" executable in such a way that these helper
functions are pre-loaded at startup.